Local/requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.3
dropbox==12.0.2
idna==3.8
numpy==1.26.4
opencv-python==4.7.0.72
ply==3.11
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.28.2
six==1.16.0
stone==3.3.1
urllib3==1.26.20
